#1220 - move modal into its own file
