+++ src/frontend/src/pages/FinancePage/FinancePage.tsx
@@ -3,7 +3,7 @@
  * See the LICENSE file in the repository root folder for details.
  */
 import { useState } from 'react';
-import { Box, FormControl, FormLabel, Grid, ListItemIcon, Menu, MenuItem, TextField } from '@mui/material';
+import { Box, Grid, ListItemIcon, Menu, MenuItem } from '@mui/material';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { NERButton } from '../../components/NERButton';
 import { useCurrentUser } from '../../hooks/users.hooks';
@@ -13,38 +13,18 @@ import ListAltIcon from '@mui/icons-material/ListAlt';
 import ReceiptIcon from '@mui/icons-material/Receipt';
 import Refunds from './RefundsSection';
 import ReimbursementRequestTable from './ReimbursementRequestsSection';
-import {
-  useAllReimbursementRequests,
-  useCurrentUserReimbursementRequests,
-  useDownloadPDFOfImages
-} from '../../hooks/finance.hooks';
+import { useAllReimbursementRequests, useCurrentUserReimbursementRequests } from '../../hooks/finance.hooks';
 import ErrorPage from '../ErrorPage';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import PageLayout from '../../components/PageLayout';
 import { useHistory } from 'react-router-dom';
 import { routes } from '../../utils/routes';
-import NERFormModal from '../../components/NERFormModal';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { Controller, useForm } from 'react-hook-form';
-import * as yup from 'yup';
-import { DatePicker } from '@mui/x-date-pickers';
-import { ReimbursementRequest } from 'shared';
-import { useToast } from '../../hooks/toasts.hooks';
-
-interface GenerateReceiptsFormInput {
-  startDate: Date;
-  endDate: Date;
-}
+import GenerateReceiptsModal from './FinanceComponents/GenerateReceiptsModal';
 
 const FinancePage = () => {
   const user = useCurrentUser();
   const history = useHistory();
-  const toast = useToast();
   const [showGenerateReceipts, setShowGenerateReceipts] = useState(false);
-  const [startDatePickerOpen, setStartDatePickerOpen] = useState(false);
-  const [endDatePickerOpen, setEndDatePickerOpen] = useState(false);
-
-  const { mutateAsync, isLoading } = useDownloadPDFOfImages();
 
   const {
     data: userReimbursementRequests,
@@ -62,43 +42,6 @@ const FinancePage = () => {
   const isFinance = user.isFinance;
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-  const schema = yup.object().shape({
-    startDate: yup.date().required('Start Date is required'),
-    endDate: yup.date().min(yup.ref('startDate'), `end date can't be before start date`).required('End Date is required')
-  });
-
-  const {
-    handleSubmit,
-    control,
-    reset,
-    formState: { errors, isValid }
-  } = useForm({
-    resolver: yupResolver(schema)
-  });
-
-  const onGenerateReceiptsSubmit = async (data: GenerateReceiptsFormInput) => {
-    if (!allReimbursementRequests) return;
-    const filteredRequests = allReimbursementRequests
-      .filter(
-        (val: ReimbursementRequest) => new Date(val.dateCreated.toDateString()) >= new Date(data.startDate.toDateString())
-      )
-      .filter(
-        (val: ReimbursementRequest) => new Date(val.dateCreated.toDateString()) <= new Date(data.endDate.toDateString())
-      )
-      .filter((val: ReimbursementRequest) => !val.dateDeleted);
-    const receipts = filteredRequests?.flatMap((request: ReimbursementRequest) => request.receiptPictures);
-    try {
-      await mutateAsync({
-        fileIds: receipts.map((receipt) => receipt.googleFileId)
-      });
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        toast.error(error.message);
-      }
-    }
-    setShowGenerateReceipts(false);
-  };
 
   if (isFinance && allReimbursementRequestsIsError) return <ErrorPage message={allReimbursementRequestsError?.message} />;
   if (userReimbursementRequestIsError) return <ErrorPage message={userReimbursementRequestError?.message} />;
@@ -158,83 +101,11 @@ const FinancePage = () => {
 
   return (
     <PageLayout title="Finance" headerRight={financeActionsDropdown}>
-      <NERFormModal
+      <GenerateReceiptsModal
         open={showGenerateReceipts}
-        onHide={() => setShowGenerateReceipts(false)}
-        title={`Generate All Receipts`}
-        reset={reset}
-        handleUseFormSubmit={handleSubmit}
-        onFormSubmit={onGenerateReceiptsSubmit}
-        formId="generate-receipts-form"
-        disabled={!isValid || isLoading}
-      >
-        {isLoading ? (
-          <LoadingIndicator />
-        ) : (
-          <>
-            <FormControl fullWidth>
-              <FormLabel>Start Date</FormLabel>
-              <Controller
-                name="startDate"
-                control={control}
-                render={({ field: { onChange, value } }) => (
-                  <DatePicker
-                    value={value}
-                    open={startDatePickerOpen}
-                    onClose={() => setStartDatePickerOpen(false)}
-                    onOpen={() => setStartDatePickerOpen(true)}
-                    onChange={(newValue) => {
-                      onChange(newValue ?? new Date());
-                    }}
-                    PopperProps={{
-                      placement: 'right'
-                    }}
-                    renderInput={(params) => (
-                      <TextField
-                        {...params}
-                        inputProps={{ ...params.inputProps, readOnly: true }}
-                        error={!!errors.startDate}
-                        helperText={errors.startDate?.message}
-                        onClick={(e) => setStartDatePickerOpen(true)}
-                      />
-                    )}
-                  />
-                )}
-              />
-            </FormControl>
-            <FormControl fullWidth>
-              <FormLabel>End Date</FormLabel>
-              <Controller
-                name="endDate"
-                control={control}
-                render={({ field: { onChange, value } }) => (
-                  <DatePicker
-                    value={value}
-                    open={endDatePickerOpen}
-                    onClose={() => setEndDatePickerOpen(false)}
-                    onOpen={() => setEndDatePickerOpen(true)}
-                    onChange={(newValue) => {
-                      onChange(newValue ?? new Date());
-                    }}
-                    PopperProps={{
-                      placement: 'right'
-                    }}
-                    renderInput={(params) => (
-                      <TextField
-                        {...params}
-                        inputProps={{ ...params.inputProps, readOnly: true }}
-                        error={!!errors.endDate}
-                        helperText={errors.endDate?.message}
-                        onClick={(e) => setEndDatePickerOpen(true)}
-                      />
-                    )}
-                  />
-                )}
-              />
-            </FormControl>
-          </>
-        )}
-      </NERFormModal>
+        setOpen={setShowGenerateReceipts}
+        allReimbursementRequests={allReimbursementRequests}
+      />
       <Grid container>
         <Grid item xs={12} sm={12} md={4}>
           <Refunds
